Travel data persistence, butter knife setup

+++ Sources/Android/PWMDCore/src/com/jraska/pwdm/core/gps/Position.java
@@ -39,7 +39,7 @@ public class Position extends LatLng implements IEventArgs
 		super.writeToParcel(dest, flags);
 
 		dest.writeLong(time);
-		dest.writeDouble(accuracy);
+		dest.writeFloat(accuracy);
 		dest.writeString(provider);
 	}
